Refactored code.| #1310
